[Numerics] Extract base class for general steady-state systems
